refactor(ZNTA-1865): Move DraggableVersionPanels and SelectableDropdown
Move DraggableVersionPanels and SelectableDropdown components to their
own folders
