+++ server/zanata-frontend/src/frontend/app/components/SelectableDropdown/index.js
