+++ server/zanata-frontend/src/frontend/app/components/index.js
@@ -14,3 +14,5 @@ export Modal from './Modal'
 export Select from './Select'
 export Sidebar from './Sidebar'
 export LockIcon from './LockIcon'
+export DraggableVersionPanels from './DraggableVersionPanels'
+export SelectableDropdown from './SelectableDropdown'
