+++ server/zanata-frontend/src/frontend/app/containers/ProjectVersion/TMMergeModal.js
@@ -4,10 +4,10 @@ import {connect} from 'react-redux'
 import { differenceWith, isEqual, throttle } from 'lodash'
 import {arrayMove} from 'react-sortable-hoc'
 import {Button, Panel, Row, InputGroup, Col, FormControl} from 'react-bootstrap'
-import {Icon, Modal, LoaderText} from '../../components'
+import {
+  Icon, Modal, LoaderText, SelectableDropdown, DraggableVersionPanels
+} from '../../components'
 import ProjectVersionPanels from './ProjectVersionPanels'
-import DraggableVersionPanels from '../../components/DraggableVersionPanels'
-import SelectableDropdown from '../../components/SelectableDropdown'
 import {ProjectVersionVertical} from './project-version-displays'
 import {ProjectVersionOptions} from './ProjectVersionOptions'
 import CancellableProgressBar
